Take care of a gcc warnign about shufting two far
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@7287 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/keytab/file/ser_ktf.c
@@ -297,7 +297,7 @@ krb5_ktf_keytab_internalize(kcontext, argp, buffer, lenremain)
 #if	SIZEOF_LONG == 4
 				    fpos = foffbuf[0];
 #else	/* SIZEOF_LONG == 4 */
-				    fpos = foffbuf[0] | (foffbuf[1] << 32);
+				    fpos = foffbuf[0] | ((long) foffbuf[1] << 32);
 #endif	/* SIZEOF_LONG == 4 */
 				    fseek(KTFILEP(keytab), fpos, SEEK_SET);
 				}
